cleaned up services
