+++ sublease-app/src/app/app-routing.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import {Routes, RouterModule, Router} from '@angular/router';
 import { AuthenticateComponent } from './Components/Forms/authenticate/authenticate.component';
 import { HomePageComponent } from './Components/home-page/home-page.component';
-import { LoginComponent } from './Components/login/login.component';
+import { LoginComponent } from './Components/forms/login/login.component';
 
 
 const routes: Routes = [
